remove unused states
