+++ template/src/utils/useSidePanel.tsx
@@ -10,7 +10,7 @@
 *********************************************
 */
 
-import React, {SetStateAction} from 'react';
+import React, {useState, SetStateAction} from 'react';
 import {SidePanelType} from '../subComponents/SidePanelEnum';
 import {createHook} from 'customization-implementation';
 
@@ -25,12 +25,14 @@ const SidePanelContext = React.createContext<SidePanelContextInterface>({
 });
 
 interface SidePanelProviderProps {
-  value: SidePanelContextInterface;
   children: React.ReactNode;
 }
 const SidePanelProvider = (props: SidePanelProviderProps) => {
+  const [sidePanel, setSidePanel] = useState<SidePanelType>(SidePanelType.None);
+
+  const value = {sidePanel, setSidePanel};
   return (
-    <SidePanelContext.Provider value={{...props.value}}>
+    <SidePanelContext.Provider value={value}>
       {props.children}
     </SidePanelContext.Provider>
   );
